+++ lib/segment/src/segment/read_view/mod.rs
@@ -4,6 +4,7 @@ mod order_by;
 mod payload;
 mod sampling;
 mod scroll;
+mod search;
 mod segment_ops;
 mod vectors;
